Add functionality do the delete button
At this point the delete button is working, if in a rough state. The way
I managed to implement it is a function within each book object. Since
the project is small, there'll be few books, and performance won't be an
issue, I won't try for more refined implementations.
At any rate, the basics of what the exercise asks is completed: there's
a button to add new books, you can click on the books after they're
added to change their read status and upon hover, you get a button to
delete entries.
TODO: there are still a few things I need to focus on now. The basics of
what the page does is there, but:
a- The popup that lets you add new books accepts anything. You can add
blank entries. TOP PRIORITY FIX
b- Even if the delete code is working, upon adding a new extry the code
that creates the default 4 books fires again, filling the table again.
c- When you reach the last book to delete, it is hard to keep the delete
button on screen because the row on which the hover listener is applied
becomes very thin and has nothing underneath. LEAST PRIORITY FIX
d- Styling fixes.